pyproject.toml
[tool.poetry]
name = "book_shop_drf_p23"
version = "0.1.0"
description = ""
authors = ["abdulloh <abdullohbaxtiyorov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
celery = "^5.4.0"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.7.0"
django-axes = "^7.0.0"
faker = "^30.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
